Add two early energy setups
One using wood pellets for fuel. The other using the liquifier to turn
crushed coal into coke instead of using the oven. The idea was to turn
the sulfuric waste water into mineralized water to help supply the algae
farms since the electrolysers for the slag for the mineralized water
consume a lot of power. Turns out this is slightly less power efficient.
You do get some practically free sulfur though but since the metal
processing loop yields slightly more sulfur than you put in this is
probably not useful.